# src/comment/dto/create-comment.dto.ts
import { ApiProperty } from '@nestjs/swagger'
import { IsNumber, IsString } from 'class-validator'
export class CreateCommentDto {
  @IsString()
  @ApiProperty()
  content: string

  @IsNumber()
  @ApiProperty()
  blogId: number

}